+++ src/main/kotlin/sonique/intellij/SoniqueDictionary.kt
@@ -1,0 +1,7 @@
+package sonique.intellij
+
+import com.intellij.spellchecker.BundledDictionaryProvider
+
+class SoniqueDictionary : BundledDictionaryProvider {
+    override fun getBundledDictionaries() = arrayOf("/dictionary/sonique.dic")
+}